fix #114 : 버튼 버그 수정
