+++ src/components/detailPage/StoreNotice.tsx
@@ -10,6 +10,7 @@ import location from '@/public/assets/icon/location.svg';
 import time from '@/public/assets/icon/timer.svg';
 import arrow from '@/public/assets/images/arrow.png';
 import {
+  Application,
   ButtonProps,
   ModalIcon,
   StoreNoticeProps,
@@ -37,8 +38,7 @@ function StoreNotice({
   isLogin,
   isProfile,
   userType,
-  applicationId,
-  setApplicationId,
+  userid,
 }: StoreNoticeProps) {
   const { hourlyPay, startsAt, workhour, description, closed } = storeData.item; //description은 이름이 겹쳐서 공고 description만 변수선언
   const { category, name, imageUrl, address1, originalHourlyPay } =
@@ -48,6 +48,7 @@ function StoreNotice({
   const startTime = formatDate(startsAt);
   const endTime = calculateEndTime(startsAt, workhour);
   const [isApplied, setIsApplied] = useState<boolean>(false); // 신청하기 버튼 상태관리변수
+  const [applicationId, setApplicationId] = useState<string>(''); // applicaionId를 담는 변수, 신청하기 버튼을 누르거나 화면이 제일 처음 렌더링될 때 값이 담김
 
   ///// 모달창과 관련된 변수들
   const [isModalOpen, setIsModalOpen] = useState(false); // 모달창 상태여부
@@ -174,12 +175,48 @@ function StoreNotice({
     }
   };
 
-  useEffect(() => {
-    // applicationId에 값이 ''이 아니면 현재 서버에 신청을 해놓은 상태니깐 버튼박스를 취소하기로 바꾼다
-    if (applicationId !== '') {
-      setIsApplied(true);
+  ///// 유저가 지원한 공고 확인하기
+  const getUserApplications = async (userId: string) => {
+    try {
+      const res = await instance.get(`/users/${userId}/applications`, {
+        headers: {
+          Authorization: `Bearer ${Cookies.get('token')}`,
+        },
+      });
+
+      const userApplications: Application[] = res.data.items; // response로 유저가 지원한 공고를 가져와서 저장(배열로 되어있음)
+      let applicationId: string = ''; // response로 받은 공고들 중에 조건에 만족하는 applicationId를 저장하기 위해 선언함
+
+      for (let i = 0; i < userApplications.length; i++) {
+        const application = userApplications[i].item;
+        if (
+          application.notice.item.id === noticeid &&
+          application.status === 'pending'
+        ) {
+          applicationId = application.id; // 위에 조건을 해당하면 그 데이터의 id를 저장하는데 id가 applicationId라고 보면 됨
+          break; // 원하는 지원을 찾았으므로 반복문 종료, 하나만 존재할건데 그래도 혹시나 싶어서 break 줬음
+        }
+      }
+      if (applicationId) {
+        // applicationId이 있다는거는 post 요청을 서버로 보냈다는 뜻이니깐 버튼상태를 true로 반환하고 applicationId을 렌더링될 때 넣어줌
+        // 이렇게 applicationId를 넣어주면 취소하기 클릭할 때 axios 에러가 생기지 않는다.
+        setApplicationId(applicationId);
+        setIsApplied(true);
+      } else {
+        setApplicationId('');
+        setIsApplied(false);
+      }
+    } catch (error) {
+      console.log('application GET 에러 : ', error);
     }
-  }, [applicationId]);
+  };
+
+  useEffect(() => {
+    if (!userid) return;
+    // getUserApplications함수가 userid가 있어야 유저가 지원했는지 확인할 수 있기 때문에 처음에 렌더링될 때 userid가 빈값인 것을 고려해서 return을 했음
+
+    getUserApplications(userid);
+  }, [userid, noticeid]);
 
   return (
     <>
